+++ CHANGELOG.md
@@ -11,12 +11,14 @@ BUG FIXES:
 * vpc: fix ignoring description when changing at inline rules of `yandex_vpc_security_group`
 * mdb: ensure that `network_id` is not empty if provided
 * mdb: fix host creating with empty `subnet_id` in `yandex_mdb_postgresql_cluster`
+* mdb: suppress master nodes config changes if no master node used in `yandex_mdb_elasticsearch_cluster`
 
 ENHANCEMENTS:
 * add `deletion_protection` attribute to `yandex_mdb_clickhouse_cluster`, `yandex_mdb_elasticsearch_cluster`, `yandex_mdb_kafka_cluster`, `yandex_mdb_mongodb_cluster`, `yandex_mdb_mysql_cluster`, `yandex_mdb_postgresql_cluster`,  `yandex_mdb_redis_cluster`, `yandex_mdb_sqlserver_cluster` resources and data sources
 * add `deletion_protection` attribute to `yandex_dataproc_cluster` data source
 * support updating of k8s node group `node_labels` without recreation.
 * docs: added a description of the port to the documentation for `yandex_mdb_postgresql_cluster`
+* support version updaing for `yandex_mdb_elasticsearch_cluster`
 
 ## 0.61.0 (July 9, 2021)
 FEATURES: